requirements.txt
django
djangorestframework
djongo